HQD-82: Falling Codeception/PlayWright CI job due problem with SmartRedirect
